Rename 'Postion' to 'Position'
This is fixing all occurrences of 'Postion' to 'Position', which looks like a typo. Unfortunately, this is changing the name of a public field, so existing users of the library will need to fix their usages when they upgrade. It's probably better to fix such a typo sooner rather than later!